fix: add return to request function and change url format
Before this fix, the login function cannot work properly. The '/' in the
request url is removed, since the separator will be provided by the url
path parameter, e.g. /auth/login. The login function also work after
adding return to the post util function to return the promise to the
caller.