make changes to build release apk
The issue faced during the build is "duplicate resource"
This issue happens when there are same resources or files with the same
name.
This issue is solved after deleting all fonts. Those fonts are most
probably added in react native also, so there is no need to add them
again during gradlew build.
Next issue would be the assets in the react native project. Although
Home.jpg and home.jpg seems different, but when react native compiles
and detects file with the same name (regardless of the case) it will
take it as duplicate resource. Therefore, we need to rename it.