change error display implementation
Previously, the error is shown to the user in a modal.
However, there is an issue of the modal overlapping all components,
making the whole page not interactable when the modal is showing.

Now, we are using the react native Alert API to show the error message.

Solve the issue https://trello.com/c/PuGJSG1m